Added postfix conversion and evaluation
Added functions to convert a provided infox string into postfix
notation, and a function to evaluate a postfix equatoin for a given set
of variables.

The equation is passed in as a string with variables in it, so I need to
make it so that the function converts variables to integers before
adding them to the stack.